Move stubs into one file to avoid import issues
