+++ src/ophyd_async/planstubs/__init__.py
@@ -1,11 +1,11 @@
-from .fly_and_collect import fly_and_collect
-from .ensure_connected import ensure_connected
-from .prepare_trigger_and_dets import (
+from .fly import (
+    fly_and_collect,
     prepare_static_seq_table_flyer_and_detectors_with_same_trigger,
 )
+from .ensure_connected import ensure_connected
 
 __all__ = [
-    "prepare_static_seq_table_flyer_and_detectors_with_same_trigger",
     "fly_and_collect",
+    "prepare_static_seq_table_flyer_and_detectors_with_same_trigger",
     "ensure_connected",
 ]
